// src/components/pages/Report/ReportBmppGroup/DetailPage/components/Detail/Detail.types.ts
export type DetailProps = {
  data?: {
    groupId?: string;
    groupName?: string;
    groupType?: string;
    groupRelation?: string;
    modal?: string;
    totalFasilitasExistingGroup?: string;
    totalFasilitasUsulanGroup?: string;
    kelonggaranBmppGroup?: string;
    persyaratanBmppGroup?: string;
    presentaseRealisasiBmppGroup?: string;
    kesimpulan?: string;
    informasiPenarikanDataModalPer?: string;
    informasiTanggalInputDataModal?: string;
    informasiFasilitasExisting?: string;
    informasiUsulanFasilitas?: string;
    dataBmppBy?: string;
    createdDate?: string;
  };
  isLoading: boolean;
};